compiled and updated readme
