+++ .gitignore
@@ -2,5 +2,5 @@
 /.vscode
 /__pycache__
 /build
-/dist/jsontranslator
-app.spec
+/venv
+jsontranslator.spec
