--- i18n/en.json
@@ -1,10 +1,0 @@
-{
-  "FirstName" : "Given Name",
-  "MiddleName" :  "Middle Name",
-  "FamilyName" :  "Last Name",
-  "Age" :  "Age",
-  "City" :  "City",
-  "Gender": "Gender",
-  "SaveRecord":"Save",
-  "CancelRecord":"Cancel"
-}
