+++ requirements.txt
@@ -1,0 +1,8 @@
+altgraph==0.17
+future==0.18.2
+json-flatten==0.1
+pefile==2019.4.18
+pyinstaller==4.1
+pyinstaller-hooks-contrib==2020.11
+PySimpleGUI==4.32.1
+pywin32-ctypes==0.2.0
